Refactore Jarvis core
* Split logic <-> UI-frontend
* Create and extract language parser abstraction
* Rename PluginManager.py, Jarvis.py, CmdInterpreter.py to follow PEP8
