+++ .gitignore
@@ -19,3 +19,5 @@ jarvis
 jarviscli/data/key_timein.json
 myinfo\.csv
 .vscode
+.buildozer
+.mypy_cache
