+++ jarviscli/__main__.py
@@ -1,7 +1,11 @@
 # -*- coding: utf-8 -*-
-import Jarvis
-import colorama
 import sys
+
+import colorama
+
+import ui.CmdInterpreter
+from jarvis import Jarvis
+from language import default
 
 
 def check_python_version():
@@ -9,12 +13,12 @@ def check_python_version():
 
 
 def main():
-    # enable color on windows
-    colorama.init()
-    # start Jarvis
-    jarvis = Jarvis.Jarvis()
+    language_parser = default.DefaultLanguageParser()
+    jarvis = Jarvis(language_parser)
+    cmd_interpreter = ui.CmdInterpreter.CmdInterpreter(jarvis)
+
     command = " ".join(sys.argv[1:]).strip()
-    jarvis.executor(command)
+    cmd_interpreter.executor(command)
 
 
 if __name__ == '__main__':
