+++ jarviscli/plugins/open.py
@@ -1,5 +1,6 @@
 import os
-from plugin import plugin, require, LINUX
+
+from plugin import LINUX, plugin, require
 
 
 @require(platform=LINUX)
@@ -11,4 +12,4 @@ def open(jarvis, s):
         os.system(string)
     else:
         jarvis.say("avaliable options are:")
-        os.system("ls -1 /usr/share/applications/ | sed -e 's/\.desktop$//'")
+        os.system("ls -1 /usr/share/applications/ | sed -e 's/\\.desktop$//'")
