+++ jarviscli/tests/test_currencyconv.py
@@ -1,7 +1,6 @@
 import unittest
-from Jarvis import Jarvis
-from plugins.currencyconv import Currencyconv
 
+from plugins.currencyconv import Currencyconv
 from tests import PluginTest
