+++ jarviscli/tests/test_evaluator.py
@@ -1,12 +1,6 @@
 import unittest
-try:  # for python2
-    from StringIO import StringIO
-except ImportError:  # for python3
-    from io import StringIO
 
-from Jarvis import Jarvis
 from plugins.evaluator import calculate
-
 from tests import PluginTest
