+++ jarviscli/tests/test_parser.py
@@ -1,42 +1,53 @@
 import unittest
-from Jarvis import Jarvis
+import language.default
+from plugin import plugin
+
+
+def do_nothing():
+    pass
 
 
 class ParserTest(unittest.TestCase):
+    def test_default(self):
+        self._test(language.default.DefaultLanguageParser())
 
-    def setUp(self):
-        self.jarvis = Jarvis()
+    def _test(self, language_parser):
+        plugins = [
+            plugin("joke")(do_nothing)(),
+            plugin("trash")(do_nothing)(),
+            plugin("test")(do_nothing)(),
+            plugin("weather")(do_nothing)(),
+            plugin("check")(do_nothing)(),
+            plugin("goodbye")(do_nothing)(),
+            plugin("say")(do_nothing)(),
+            plugin("near")(do_nothing)()
+        ]
 
-    def test_chuck(self):
+        language_parser.train(plugins)
+
         user_input = "Jarvis, I want to hear a joke about Chuck Norris, can you help me?"
-        parsed_input = self.jarvis.parse_input(user_input).split()
-        self.assertEqual("joke", parsed_input[0])
+        parsed_input = language_parser.identify_action(user_input)
+        self.assertEqual("joke", parsed_input.get_name())
 
-    def test_weather(self):
         user_input = "Mmm... I want to go for a walk. What's the weather like today?"
-        parsed_input = self.jarvis.parse_input(user_input).split()
-        self.assertEqual("weather", parsed_input[0])
+        parsed_input = language_parser.identify_action(user_input)
+        self.assertEqual("weather", parsed_input.get_name())
 
-    def test_goodbye(self):
         user_input = "Thanks for your hard work Jarvis, goodbye!"
-        parsed_input = self.jarvis.parse_input(user_input).split()
-        self.assertEqual("goodbye", parsed_input[0])
+        parsed_input = language_parser.identify_action(user_input)
+        self.assertEqual("goodbye", parsed_input.get_name())
 
-    def test_check_ram(self):
         user_input = "It would be cool if you could check my computers ram"
-        parsed_input = self.jarvis.parse_input(user_input).split()
-        self.assertEqual(["check"], parsed_input[0:1])
+        parsed_input = language_parser.identify_action(user_input)
+        self.assertEqual("check", parsed_input.get_name())
 
-    def test_say(self):
         user_input = "Can you say I'm a robot"
-        parsed_input = self.jarvis.parse_input(user_input).split()
-        self.assertEqual(["say", "i'm", "a", "robot"], parsed_input[0:])
+        parsed_input = language_parser.identify_action(user_input)
+        self.assertEqual("say", parsed_input.get_name())
 
-    def test_near(self):
         user_input = "charities near Valencia"
-        parsed_input = self.jarvis.parse_input(user_input).split()
-        self.assertEqual(
-            ["near", "charities", "|", "valencia"], parsed_input[0:])
+        parsed_input = language_parser.identify_action(user_input)
+        self.assertEqual("near", parsed_input.get_name())
 
 
 if __name__ == '__main__':
